+++ packages/arcgis-rest-portal/src/groups/helpers.ts
@@ -1,12 +1,12 @@
 /* Copyright (c) 2017-2019 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import { IUserRequestOptions } from "@esri/arcgis-rest-request";
+import { IAuthenticatedRequestOptions } from "@esri/arcgis-rest-request";
 
 /**
  * Base options interface for making authenticated requests for groups.
  */
-export interface IUserGroupOptions extends IUserRequestOptions {
+export interface IUserGroupOptions extends IAuthenticatedRequestOptions {
   /**
    * Unique identifier of the group.
    */