+++ packages/arcgis-rest-portal/src/users/set-user-properties.ts
@@ -1,7 +1,10 @@
 /* Copyright (c) 2023 Environmental Systems Research Institute, Inc.
  * Apache-2.0 */
 
-import { IUserRequestOptions, request } from "@esri/arcgis-rest-request";
+import {
+  IAuthenticatedRequestOptions,
+  request
+} from "@esri/arcgis-rest-request";
 import { getPortalUrl } from "../util/get-portal-url.js";
 import { IUserProperties } from "./get-user-properties.js";
 
@@ -9,18 +12,18 @@ import { IUserProperties } from "./get-user-properties.js";
  * Updates the properties for a user
  * @param username The user whose properties to update
  * @param properties IUserProperties object with properties to update
- * @param requestOptions An IUserRequestOptions object
+ * @param requestOptions An IAuthenticatedRequestOptions object
  * @returns a promise that resolves to { success: boolean }
  */
 export async function setUserProperties(
   username: string,
   properties: IUserProperties,
-  requestOptions: IUserRequestOptions
+  requestOptions: IAuthenticatedRequestOptions
 ): Promise<{ success: boolean }> {
   const url = `${getPortalUrl(
     requestOptions
   )}/community/users/${encodeURIComponent(username)}/setProperties`;
-  const options: IUserRequestOptions = {
+  const options: IAuthenticatedRequestOptions = {
     httpMethod: "POST",
     params: { properties },
     ...requestOptions